schema: split schema package into multiple files
